Make IncomingBox its own component
