+++ assets/css/_route_ladder.scss
@@ -36,30 +36,3 @@
     fill: $charcoal-grey;
   }
 }
-
-.m-incoming-box {
-  align-items: center;
-  background-color: $lightest-grey;
-  color: $dark;
-  display: flex;
-  flex-direction: column;
-  font-size: 0.8125rem;
-  min-height: 2.5rem;
-  padding: 0.5rem;
-  width: 12em;
-}
-
-.m-incoming-box__vehicle {
-  cursor: pointer;
-  display: flex;
-  margin-bottom: 0.125rem;
-}
-
-.m-incoming-box__vehicle-icon {
-  margin-right: 0.5rem;
-}
-
-.m-incoming-box__vehicle-icon-svg {
-  height: 14px;
-  width: 18px;
-}
