+++ assets/css/app.scss
@@ -3,6 +3,7 @@
 @import "inter_ui";
 
 @import "close_button";
+@import "incoming_box";
 @import "ladder";
 @import "route_ladder";
 @import "route_ladders";
