+++ assets/src/components/routeLadder.tsx
@@ -2,16 +2,11 @@ import React, { useContext, useState } from "react"
 import DispatchContext from "../contexts/dispatchContext"
 import { reverseIcon, reverseIconReversed } from "../helpers/icon"
 import { LoadableTimepoints, Route, Vehicle, VehicleId } from "../skate"
-import { deselectRoute, selectVehicle } from "../state"
+import { deselectRoute } from "../state"
 import CloseButton from "./closeButton"
-import Ladder, {
-  flipLadderDirection,
-  LadderDirection,
-  VehicleDirection,
-  vehicleDirectionOnLadder,
-} from "./ladder"
+import IncomingBox from "./incomingBox"
+import Ladder, { flipLadderDirection, LadderDirection } from "./ladder"
 import Loading from "./loading"
-import VehicleIcon from "./vehicleIcon"
 
 interface Props {
   route: Route
@@ -40,60 +35,6 @@ const Header = ({ route }: { route: Route }) => {
     </div>
   )
 }
-
-const IncomingBoxVehicle = ({
-  vehicle,
-  ladderDirection,
-  selectedVehicleId,
-}: {
-  vehicle: Vehicle
-  ladderDirection: LadderDirection
-  selectedVehicleId: VehicleId | undefined
-}) => {
-  const dispatch = useContext(DispatchContext)
-  const isSelected = vehicle.id === selectedVehicleId
-  const rotation =
-    vehicleDirectionOnLadder(vehicle, ladderDirection) === VehicleDirection.Down
-      ? 180
-      : 0
-
-  return (
-    <div
-      className="m-incoming-box__vehicle"
-      onClick={() => dispatch(selectVehicle(vehicle.id))}
-    >
-      <div className="m-incoming-box__vehicle-icon">
-        <svg className="m-incoming-box__vehicle-icon-svg">
-          <g transform={`rotate(${rotation},9,7)`}>
-            <VehicleIcon isSelected={isSelected} scale={0.38} />
-          </g>
-        </svg>
-      </div>
-      <div className="m-incoming-box__vehicle-label">{vehicle.label}</div>
-    </div>
-  )
-}
-
-const IncomingBox = ({
-  vehicles,
-  ladderDirection,
-  selectedVehicleId,
-}: {
-  vehicles: Vehicle[]
-  ladderDirection: LadderDirection
-  selectedVehicleId: VehicleId | undefined
-}) => (
-  <div className="m-incoming-box">
-    {vehicles.map(vehicle => (
-      <IncomingBoxVehicle
-        vehicle={vehicle}
-        ladderDirection={ladderDirection}
-        selectedVehicleId={selectedVehicleId}
-        key={vehicle.id}
-      />
-    ))}
-  </div>
-)
 
 const RouteLadder = ({
   route,
